fix: address code review findings #1, #4, #5
#1 Remove generated files committed by hand
   ESPHOME-release/ is generated output that must only change via a full
   prepare-component-release run. These two files leaked in from a stray
   partial regeneration and are dead code in the ESPHome component
   (nothing there calls radian_decode_4bitpbit).
     - ESPHOME-release/everblu_meter/radian_decoder.cpp
     - ESPHOME-release/everblu_meter/radian_decoder.h

#4 Replace Unicode in tool output (tools/hex_decoder.cpp)
   The CRC line still used the check/cross glyphs while everything else
   had been ASCII-fied. These render as mojibake in Windows consoles for
   full 124-byte frames. Now "VALID" / "INVALID (CRC mismatch)".

#5 Fix narrowing cast in decoder bounds checks (src/core/radian_decoder.cpp)
   Casting decoded_max (int) to uint8_t wrapped for values > 255,
   defeating the bounds guard. dest_byte_cnt is already uint8_t, so it
   promotes to int and compares safely against decoded_max directly."

+++ src/core/radian_decoder.cpp
@@ -69,7 +69,7 @@ uint8_t radian_decode_4bitpbit(const uint8_t *rx_buf, int rx_len,
                     if (dest_bit_cnt < 8)
                     {
                         /* Data bit: shift in from MSB */
-                        if (dest_byte_cnt >= (uint8_t)decoded_max)
+                        if (dest_byte_cnt >= decoded_max)
                             return dest_byte_cnt;
 
                         decoded[dest_byte_cnt] >>= 1;
@@ -86,7 +86,7 @@ uint8_t radian_decode_4bitpbit(const uint8_t *rx_buf, int rx_len,
                             framing_errors++;
                         dest_bit_cnt = 0;
                         dest_byte_cnt++;
-                        if (dest_byte_cnt >= (uint8_t)decoded_max)
+                        if (dest_byte_cnt >= decoded_max)
                             return dest_byte_cnt;
                         continue;
                     }
@@ -96,7 +96,7 @@ uint8_t radian_decode_4bitpbit(const uint8_t *rx_buf, int rx_len,
                     {
                         dest_bit_cnt = 0;
                         dest_byte_cnt++;
-                        if (dest_byte_cnt >= (uint8_t)decoded_max)
+                        if (dest_byte_cnt >= decoded_max)
                             return dest_byte_cnt;
                     }
                 }
